refactor: Logger format utils
