+++ src/Utils/Logger/Format/RemoveError.js
@@ -1,0 +1,7 @@
+import winston from 'winston'
+
+const RemoveError = winston.format(info => {
+	return info.level !== 'error' ? info : false
+})
+
+export default RemoveError
