+++ src/Utils/Logger/Format/RemoveHttp.js
@@ -1,0 +1,7 @@
+import winston from 'winston'
+
+const RemoveHttp = winston.format(info => {
+	return info.level !== 'http' ? info : false
+})
+
+export default RemoveHttp
