+++ src/Utils/Logger/Format/RemoveInfo.js
@@ -1,0 +1,7 @@
+import winston from 'winston'
+
+const RemoveInfo = winston.format(info => {
+	return info.level !== 'info' ? info : false
+})
+
+export default RemoveInfo
