+++ src/Utils/Logger/Format/RemoveWarn.js
@@ -1,0 +1,7 @@
+import winston from 'winston'
+
+const RemoveWarn = winston.format(info => {
+	return info.level !== 'warn' ? info : false
+})
+
+export default RemoveWarn
